fix(optimx): Fix version checking for optimx
It appears the behavior for the line

```
  if(packageVersion("optimx") == 2017)
```

has changed such that the right hand side is coerced into a version
string via (after some other calls) `make_numeric_version`, which only accepts
character strings. The constant `2017` is interpreted as a double, which
causes the function fail, which halts the script. The solution is simply
to turn the numeric constants to string constants in this context.

Additionally, I replaced a unicode dash with an ascii dash, as it was
causing problems when sourcing the script.